==> notebooks/00_read_ontology_files.ipynb <==
!pip install rdflib

from google.colab import drive
drive.mount('/content/drive')

import rdflib
from rdflib import Graph, RDFS, RDF, URIRef
import requests
import pandas as pd
import sys
sys.path.append('/content/drive/MyDrive/Projects/semantic_alignment_ontology/src')
import ontology_utils
import importlib
importlib.reload(ontology_utils)
from ontology_utils import load_ontology_labels

# Read ontology data: PTO (Plant Trait Ontology)
url = "http://purl.obolibrary.org/obo/to.owl"
headers = {"Accept": "application/rdf+xml"}
response = requests.get(url, headers=headers)

with open("/content/drive/MyDrive/Projects/semantic_alignment_ontology/data/to.owl", "wb") as f:
  f.write(response.content)

df_to = load_ontology_labels("/content/drive/MyDrive/Projects/semantic_alignment_ontology/data/to.owl", format="xml")

print(df_to)

# Read ontology data: GO (Gene Ontology)
url = "https://purl.obolibrary.org/obo/go.owl"
headers = {"Accept": "application/rdf+xml"}
r = requests.get(url, headers=headers)

with open("/content/drive/MyDrive/Projects/semantic_alignment_ontology/data/go.owl", "wb") as f:
  f.write(r.content)

df_go = load_ontology_labels("/content/drive/MyDrive/Projects/semantic_alignment_ontology/data/go.owl", format="xml")

print(df_go)

# save the dataframe for later use
df_to.to_pickle("/content/drive/MyDrive/Projects/semantic_alignment_ontology/data/to.pkl")
df_go.to_pickle("/content/drive/MyDrive/Projects/semantic_alignment_ontology/data/go.pkl")
