==> notebooks/01_ontology_to_embeddings.ipynb <==
from google.colab import drive
drive.mount('/content/drive')

Step 1: Pretrained language model and data(frame) files prep for embedding computing per label

# install the language model to gnerate similarity values per label
!pip install -U sentence-transformers

import pandas as pd
from sentence_transformers import SentenceTransformer

# load the dataframe files
# N.B. each column of df contains four columns: URI, Label, Definition, Text (label+definition)
df_to = pd.read_pickle("/content/drive/MyDrive/Projects/semantic_alignment_ontology/data/to.pkl")
df_go = pd.read_pickle("/content/drive/MyDrive/Projects/semantic_alignment_ontology/data/go.pkl")

Step 2: Computing vector embeddings

# Set up the pretrained language model
model = SentenceTransformer("all-MiniLM-L6-v2")

# Compute sentence embeddings per label

# Encode the Text (Label+Definition) from Plant trait ontology
df_to["Embedding"] = df_to["Text"].apply(lambda x: model.encode(x, show_progress_bar=False))

# Also encode the Text from Gene Ontology (but had to do that step by step due to network failure)
df_go["Embedding"] = None  # initialize the column

for inx in range(45000,len(df_go["Text"])):
  text = df_go["Text"].iloc[inx]
  emb = model.encode(text, show_progress_bar=False)
  df_go.at[inx, "Embedding"] = emb

# Double check if embeddings are computed properly
import numpy as np
# shape
#print(np.array(df_to["Embedding"].iloc[0]).shape) # should be (384,) as the vector size is 384
print(np.array(df_go["Embedding"].iloc[48000])) # same
# one embedding?
#print(np.array(df_to["Text"].iloc[100]))
#print(np.array(df_to["Embedding"].iloc[100]))

# Save the df with embeddings for later
#df_to.to_pickle("/content/drive/MyDrive/Projects/semantic_alignment_ontology/data/to_embeddings.pkl")
df_go.to_pickle("/content/drive/MyDrive/Projects/semantic_alignment_ontology/data/go_embeddings.pkl")
